+++ Client/BlockProcessing/Game/Features/Terrain/Chunk/ChunkManager.h
@@ -18,9 +18,9 @@ public:
     void renderChunk(Chunk* chunk);
     void unloadChunk(Chunk* chunk);
     int8_t getChunkBlock(Chunk *chunk, int64_t x, int64_t y, int64_t z);
-    int8_t getChunkBlock(Chunk *chunk, int64_t x, int64_t y, int64_t z, int64_t indexX, int64_t indexY, int64_t indexZ);
+    int8_t getChunkBlock(Chunk *chunk, int64_t x, int64_t y, int64_t z, int8_t indexX, int8_t indexY, int8_t indexZ);
     void setChunkBlock(Chunk *chunk, int8_t block, int64_t x, int64_t y, int64_t z);
-    void setChunkBlockIndexed(Chunk *chunk, int8_t block, int x, int y, int z);
+    void setChunkBlockIndexed(Chunk *chunk, int8_t block, int8_t x, int8_t y, int8_t z);
 private:
     Mat4d model;
     Mat4d view;
@@ -28,7 +28,7 @@ private:
     int stride;
     void generateChunkVertices(Chunk* chunk);
     void generateChunkDefaultVertices(Chunk* chunk);
-    void addFace(std::vector<float>& data, Block* block, int x, int y, int z, int face, int lightLevel);
+    void addFace(std::vector<float>& data, Block* block, int8_t x, int8_t y, int8_t z, int8_t face, int8_t lightLevel);
     BlockManager *blockManager;
     WorldManager *worldManager;
 };
